fix(uta): export utaConfigSchema from uta-protocol — unbreaks test-connection
The UTA service's `/test-connection` endpoint (routes-trading.ts:191)
imports `utaConfigSchema` from `@traderalice/uta-protocol`, but the
schema was only defined in Alice's `src/core/config.ts` and never
exported from the protocol package. This caused a TypeError:

  Cannot read properties of undefined (reading 'parse')

…on every "Test <exchange>" action in the New UTA flow (affects all
exchanges, not just Bitget).

Fix: define `utaConfigSchema` + `guardConfigSchema` in the protocol
package (`packages/uta-protocol/src/schemas/index.ts`) and re-export
from `src/core/config.ts` to keep the existing public API intact.
